+++ packages/vanilla-components/src/components/rich-radio-option/config.ts
@@ -6,7 +6,7 @@ export const richRadioOptionConfig = {
   fixedClasses: {
     disabled: commonClasses.optionOrItemDisabled,
 
-    option: '',
+    option: 'select-none',
 
     checked: '',
     unchecked: '',